Optionally install gems before testing

+++ test-eventide-postgres.sh
@@ -4,9 +4,17 @@ set -e
 
 clear
 
+install_gems=false
+if [ ! -z ${INSTALL_GEMS+x} ]; then
+  if [ $INSTALL_GEMS = true ]; then
+    install_gems=true
+  fi
+fi
+
 echo
 echo "Testing Eventide Postgres Libraries"
 echo "= = ="
+echo "(Install Gems: $install_gems}"
 echo
 
 libraries=(
@@ -21,13 +29,6 @@ libraries=(
   "entity-snapshot-postgres"
 )
 
-install_gems=false
-if [ -z ${INSTALL_GEMS+x} ]; then
-  if [ $INSTALL_GEMS = true ]; then
-    install_gems=true
-  fi
-fi
-
 pushd $PROJECTS_HOME > /dev/null
 
 for name in "${libraries[@]}"; do
@@ -40,9 +41,18 @@ for name in "${libraries[@]}"; do
 
   if [ $install_gems = true ]; then
     echo "- installing gems"
-    rm -rf gems
-    rm -rf .bundle
-    rm Gemfile.lock
+
+    if [ -d gems ]; then
+      rm -rf gems
+    fi
+
+    if [ -d .bundle ]; then
+      rm -rf .bundle
+    fi
+
+    if [ -f Gemfile.lock ]; then
+      rm Gemfile.lock
+    fi
 
     . ./install-gems.sh
   fi
